src: fix clippy::module_inception for pipeline module
