+++ src/stream/pipeline/v4l_pipeline.rs
@@ -7,9 +7,7 @@ use crate::{
     video_stream::types::VideoAndStreamInformation,
 };
 
-use super::pipeline::{
-    PipelineGstreamerInterface, PipelineState, PIPELINE_FILTER_NAME, PIPELINE_TEE_NAME,
-};
+use super::{PipelineGstreamerInterface, PipelineState, PIPELINE_FILTER_NAME, PIPELINE_TEE_NAME};
 
 use anyhow::{anyhow, Context, Result};
