+++ src/stream/stream.rs
@@ -1,5 +1,5 @@
 use super::manager::Manager;
-use super::pipeline::pipeline::Pipeline;
+use super::pipeline::Pipeline;
 use super::sink::sink::{create_rtsp_sink, create_udp_sink};
 use super::types::*;
 use super::webrtc::signalling_protocol::PeerId;
